refactor: [User] AuthService 트랜잭션 분리 및 로그인 이력 기록 개선 #24
- AuthTransactionalService: 트랜잭션 DB 작업 분리 (JWT 발급, 사용자 저장)
- AuthService: 외부 API 검증 후 트랜잭션 서비스 위임
- LoginHistoryRecorder: 독립 트랜잭션(REQUIRES_NEW)으로 이력 기록
- 외부 API 실패(reCAPTCHA/PortOne) 시에도 이력 기록 보장
